+++ src/parser/enrichers/class/barbarian/DivineFury.mjs
@@ -1,0 +1,32 @@
+/* eslint-disable class-methods-use-this */
+import DDBEnricherData from "../../data/DDBEnricherData.mjs";
+
+export default class DivineFury extends DDBEnricherData {
+
+  get type() {
+    return "damage";
+  }
+
+  get activity() {
+    return {
+      targetType: "creature",
+      activationType: "special",
+      data: {
+        damage: {
+          critical: {
+            allow: true,
+            parts: [
+              DDBEnricherData.basicDamagePart({
+                bonus: "(floor(@classes.barbarian.levels / 2))",
+                number: 1,
+                denomination: 6,
+                types: ["necrotic", "radiant"],
+              }),
+            ],
+          },
+        },
+      },
+    };
+  }
+
+}